# swagger.yaml
openapi: 3.0.0
info:
  title: "Food Delivery Backend APIs"
  version: 0.1.9
  description: "This is our Food Delivery's App Server."
  version: "1.0.0"
  contact:
    email: "abhishsusu@gmail.com"
servers:
  - url: http://localhost:8000/api/v1//
    description: Locally Hosted
  - url: https://urbaan-eatery.herokuapp.com/api/v1// 
    description: Our Hosted Server
...
paths:
  food/getallfoods:
    get:  
      tags:
      - "Foods"
      summary: Returns a list of all foods.
      description: It returns us with all the foods that are present.
      operationId: "getAllFoods"
      responses:
        '200':    # status code 
          description: A JSON array of food items
          content:
            application/json:
              schema: 
                type: array
                items: 
                  type: string
        '400':
          description: "Invalid status value"
...